added Translated Text

+++ src/main/java/com/wayahead/arrivechat/repository/MessageRepository.java
@@ -6,6 +6,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.List;
 
 
-public interface MessageRepository extends JpaRepository<Messages, Long> {
-    List<Messages> findByUserId(Long userId);
+public interface MessageRepository extends JpaRepository<Messages, Integer> {
+    List<Messages> findByUserId(Integer userId);
 }
